+++ src/main/java/com/example/demo/repositories/DriverRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repositories;
 import com.example.demo.entities.Driver;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DriverRepository extends JpaRepository<Driver, Long> {
+public interface DriverRepository extends JpaRepository<Driver, Integer> {
 }
